Move footer constant types to types.ts
